+++ pyHMT2D/Misc/gmsh2d_to_srh.py
@@ -1,0 +1,322 @@
+"""
+A tool to convert Gmsh's 2D MSH file to srhgeom format (to be used by SRH-2D). Optionally, it can interpolate terrain
+to the 2D mesh. Gmsh's 2D mesh usually do not have elevation information.
+"""
+
+import numpy as np
+import sys
+import meshio
+
+
+def gmsh2d_to_srh(gmsh2d_fileName, srh_caseName, units="Meters"):
+    """Convert Gmsh 2D mesh into SRH-2D format
+
+    It generates two files: srhgeom for mesh and srhmat for Manning's n
+
+    The srhhydro file has to be generated separately.
+
+    Parameters
+    ----------
+    gmsh2d_fileName : str
+        file name of the Gmsh MSH file
+    srh_caseName : str
+        case name for SRH-2D.
+    units : str, default Meters
+        length units of Gmsh file
+
+    Returns
+    -------
+
+    """
+
+    print("Converting Gmsh's MSH mesh to SRH-2D format ...")
+
+    #read in the Gmsh MSH with meshio
+    mesh = meshio.read(gmsh2d_fileName)
+
+    #output the mesh as vtk for checking
+    mesh.write("check_mesh.vtk")
+
+    #build the nodeStrings (boundaries)
+    nodeStrings = build_nodeStrings(mesh)
+
+    #write srhgeom file
+    write_srhgeom(srh_caseName, mesh, nodeStrings, units)
+
+    #build Manning's n zone information
+    nManningNZones, ManningNZoneNames, cellsInManningZones = build_ManningNZones(mesh)
+
+    #write srhmat file
+    write_srhmat(srh_caseName, nManningNZones, ManningNZoneNames, cellsInManningZones)
+
+    print("Finished converting Gmsh's MSH mesh to SRH-2D format.")
+
+
+def build_nodeStrings(mesh):
+    """Build the nodeStrings (boundary lines)
+
+    Parameters
+    ----------
+    mesh : meshio's mesh
+        meshio's mesh object
+
+    Returns
+    -------
+
+    """
+
+    #get all the lines's physcial group number
+    lines_physical_group_IDs = mesh.cell_data_dict['gmsh:physical']['line']
+
+    #how many different line boundaries
+    unique_line_physical_group_IDs = np.unique(lines_physical_group_IDs)
+
+    #get all lines' nodes (assume there is only one line list. Sure?)
+    lines_nodes = mesh.cells[0].data
+
+    #build node strings for each line boundary
+    nodeStrings = {}
+
+    for nodeStringID in unique_line_physical_group_IDs:
+        #print("nodeStringID =", nodeStringID)
+
+        current_node_list = []
+
+        #loop through each line in lines_physical_group_IDs:
+        for line_group_ID, line_ID in zip(lines_physical_group_IDs, range(len(lines_physical_group_IDs))):
+            #print("line_group_ID, line_ID =", line_group_ID, line_ID)
+
+            if nodeStringID == line_group_ID:
+                # add the nodes of line to current_node_list (if they are not in already)
+                if (lines_nodes[line_ID][0]+1) not in current_node_list:
+                    current_node_list.append(lines_nodes[line_ID][0]+1)  #+1 because meshio is 0-based; gmsh is 1-based
+
+                if (lines_nodes[line_ID][1]+1) not in current_node_list:
+                    current_node_list.append(lines_nodes[line_ID][1]+1)  #+1 because meshio is 0-based; gmsh is 1-based
+
+        #add the nodeString to the nodeStrings dictionary
+        nodeStrings[nodeStringID] = current_node_list
+
+    #print(type(nodeStrings))
+
+    return nodeStrings
+
+def build_ManningNZones(mesh):
[…]
