+++ pyHMT2D/Misc/__init__.py
@@ -2,5 +2,6 @@ from .tools import *
 from .Terrain import *
 from .RAS_to_SRH_Converter import *
 from .vtk_utilities import *
+from .gmsh2d_to_srh import *
 
-__all__ = ["tools", "Terrain", "RAS_to_SRH_Converter", "vtk_utilities"]
+__all__ = ["tools", "Terrain", "RAS_to_SRH_Converter", "vtk_utilities", "gmsh2d_to_srh"]
