+++ pyHMT2D/cli/__init__.py
@@ -1,4 +1,5 @@
 from .ras_to_srh import ras_to_srh
 from .hmt_calibrate import hmt_calibrate
+from .srh_mesh_to_vtk import hmt_srh_mesh_to_vtk
 
 __all__ = ["ras_to_srh", "hmt_calibrate"]

[…]
