+++ pyHMT2D/cli/__init__.py
@@ -1,5 +1,6 @@
 from .ras_to_srh import ras_to_srh
 from .hmt_calibrate import hmt_calibrate
 from .srh_mesh_to_vtk import hmt_srh_mesh_to_vtk
+from .srh_to_vtk import hmt_srh_to_vtk
 
-__all__ = ["ras_to_srh", "hmt_calibrate"]
+__all__ = ["ras_to_srh", "hmt_calibrate", "hmt_srh_mesh_to_vtk", "hmt_srh_to_vtk"]
