+++ examples/calibration/Backwater-1D/backwater/demo_calibration_Backwater_1D.py
@@ -105,7 +105,8 @@ def demo_calibrator():
 
     """
 
-    my_calibrator = pyHMT2D.Calibration.Calibrator("calibration.json")
+    #my_calibrator = pyHMT2D.Calibration.Calibrator("calibration.json")
+    my_calibrator = pyHMT2D.Calibration.Calibrator("calibration_enumerator.json")
 
     my_calibrator.calibrate()
 
