### 2021-05-12 dbe2367
more specific version requirement for Python and vtk
