+++ CHANGELOG.md
@@ -3,4 +3,8 @@
 
 
 
-## v1.0.0
+## v1.0.5
+
+- added enumerator optimizer
+- enforced python version to be <3.8
+- enforced vkt version to be < 8.2
