### 2021-05-12 d14bd6d
more specific version requirement for Python and vtk
