### 2021-05-13 #3180 @kthyng
@psu-efd Thanks for the reviewer ideas!

@jmp75 and @AtrCheema — would you be interested in reviewing this submission for the Journal for Open Source Software? We would ask you to start the review in the next few weeks to be finished in 4-6 weeks, anticipated some potential back and forth of iteration with the author.  It's all done through github — you can learn more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
