### 2021-05-14 #3180 @jmp75
@kthyng  I am interested. This will be my first JOSS review but I'll try to do a suitable job.
