### 2021-05-17 9417a42
Update process_RAS_2D_Data.py
The train (constant_slope_channel.tif) is in folder: "\examples\compare_SRH_2D_RAS_2D\backwater_curves"
