### 2021-06-01 #3180 @arfon
> @openjournals/dev The paper here never seems to have any error nor any follow up from whedon, despite repeated calls to generate the pdf over time.

@kthyng – I'm not exactly sure what's going on here but compiling the paper is blowing up the memory usage of Whedon and the process is being killed.

Historically this has been due to large figures but the PNGs here don't look too big. I'll need to dig into this further.

For now though, here's the paper compiled locally: [10.21105.joss.03180.pdf](https://github.com/openjournals/joss-reviews/files/6577596/10.21105.joss.03180.pdf)
