### 2021-06-01 #3180 @mdbartos
> Hi @mdbartos! Would you be interested in reviewing this submission for the Journal for Open Source Software? We would ask you to start the review in the next few weeks to be finished in 4-6 weeks, anticipated some potential back and forth of iteration with the author. It's all done through github — you can learn more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Sure, happy to help. I've had a paper reviewed before in JOSS and I'm a big fan of the process.
