### 2021-07-13 #3332 @mdbartos
Same; just getting back from some summer travel and will complete my checklist soon.
