### 2021-07-16 psu-efd/pyHMT2D#4 @psu-efd
>
>
> Greetings,
>
> Is this a Windows-only package?
>
> When trying to import the module I get the following traceback:
>
> ```
> >>> import pyHMT2D
> Traceback (most recent call last):
>   File "<stdin>", line 1, in <module>
>   File "/usr/local/lib/python3.7/site-packages/pyHMT2D/__init__.py", line 2, in <module>
>     from .Hydraulic_Models_Data import *
>   File "/usr/local/lib/python3.7/site-packages/pyHMT2D/Hydraulic_Models_Data/__init__.py", line 4, in <module>
>     from .RAS_2D import *
>   File "/usr/local/lib/python3.7/site-packages/pyHMT2D/Hydraulic_Models_Data/RAS_2D/__init__.py", line 2, in <module>
>     from .HEC_RAS_Model import HEC_RAS_Model
>   File "/usr/local/lib/python3.7/site-packages/pyHMT2D/Hydraulic_Models_Data/RAS_2D/HEC_RAS_Model.py", line 5, in <module>
>     import win32com.client as win32
> ModuleNotFoundError: No module named 'win32com'
> ```

Thanks for testing the package. Yes, this package is intended only to work on Windows as of now because both SRH-2D and HEC-RAS are only supported in Windows. Maybe I should add this to the README.
