### 2021-07-21 #3332 @psu-efd
Hi Matt,

Thanks for reviewing this work. How about virtual machine?

Xiaofeng Liu

From: Matt Bartos ***@***.***>
Sent: Wednesday, July 21, 2021 12:47 AM
Cc: Dr. Xiaofeng Liu, Environmental Fluid Dynamics Group, Penn State University ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: pyHMT2D: Python-based two-dimensional hydraulic modeling tools (#3332)

Greetings,

Thanks for your help. Unfortunately, the link doesn't seem to work for me.

I've completed all the checklist items that I can complete. Unfortunately, I am not sure that I will be able to complete all aspects of the review at this time. The package is Windows-only (psu-efd/pyHMT2D#4 <https://github.com/psu-efd/pyHMT2D/issues/4> ), and also requires access to external software that is only supported in Windows.

Currently, I only have access to mac and linux machines, and thus can't install/test the software. I'm planning to purchase a couple of Windows machines in the near future for another project, but I'm not exactly sure about the timeline.

Perhaps it might be possible to install the package on a windows instance in the cloud?

Let me know how you'd like to proceed.

Thanks,
MDB
