### 2021-07-21 #3332 @psu-efd
>
>
> @mdbartos Do you have multiple github accounts? If so make sure you are signed into the one associated with this review when clicking on the link and working on the review. If there is still an issue I'll ask for help.
>
> @mdbartos and @psu-efd: Will a virtual machine approach work? Or something like Binder? The limitation of this being able to run in Windows only should be clear in the readme as well as the docs.

I think a virtual machine should work. Binder will not work. This package is developed to help run 2D hydraulics models, which currently support U.S. Army Corps of Engineer's HEC-RAS and U.S. Bureau of Reclamation's SRH-2D. Both only run on Windows because they are used by engineers and practitioners.

I will highlight in the README and the User Manual that currently only Windows is supported (this should be apparent for users of this Python Package though).
