### 2021-07-21 4014fca
added information about the only supported OS is Windows
