### 2021-07-21 9268b88
added information about the only supported OS is Windows
