### 2021-07-21 241e17a
added information in readme about the prerequisite of installing 2D models
