### 2021-07-21 #3332 @kthyng
The user manual currently specifically says that it works with Mac and Linux, and of course the point of a user manual is to help new users figure out how to use your software including whether they should be using it at all. Same for the JOSS paper.

I think it is a valid argument that engineers tend to use Windows, but certainly that doesn't mean the software should by definition only run on Windows if engineers use it.

@psu-efd Do you have everything set up for a virtual machine so that a user can easily set it up and run it?
