### 2021-07-21 #3332 @psu-efd
>
>
> The user manual currently specifically says that it works with Mac and Linux, and of course the point of a user manual is to help new users figure out how to use your software including whether they should be using it at all. Same for the JOSS paper.

That is a good point. I have temporally changed the README to point out that if the user wants to run 2D models on Windows, then they have to install and use pyHMT2D on Windows. There are other 2D models (currently not supported) which can run on Linux and Mac. pyHMT2D can be used on Linux and Mac because it is a python library and the python code itself can be run on any OS. That is why on Page 6, Section 2.1.1 I pointed out that "We use Windows operating system as an example. pyHMT2D can be used in Linux and Mac OS".

>
> I think it is a valid argument that engineers tend to use Windows, but certainly that doesn't mean the software should by definition only run on Windows if engineers use it.
>
> @psu-efd Do you have everything set up for a virtual machine so that a user can easily set it up and run it?

Virtual machine should work. If anyone wants to try VM, Windows should be installed first and then HEC-RAS and SRH-2D. The rest is identical to a dedicated Windows machine. Another option is to use Cloud service (select a Windows instance).

I hope my answer is to your satisfaction. Thanks.
