### 2021-07-21 #3332 @kthyng
@psu-efd For the review process, the reviewers need to actually run your software. So, either you need to make it really easy for @mdbartos to do this with a virtual machine setup or another approach (by giving clear step-by-step instructions, etc), or we need to add another reviewer who can run the windows-based software. Please remember that the reviewers are volunteering their time to improve your software through this review process, so we need to make the mechanisms for that as easy and fast as we can. Additionally, you may have future users who would want to use your software but have the same problem.
