### 2021-07-21 #3332 @mdbartos
Another solution could be to include pre-compiled binaries for different platforms, like PySWMM does with the EPA stormwater management model (https://github.com/OpenWaterAnalytics/pyswmm).

Not sure about whether USBR or USACE provide access to source code though.
