### 2021-07-21 #3332 @psu-efd
>
>
> Another solution could be to include pre-compiled binaries for different platforms, like PySWMM does with the EPA stormwater management model (https://github.com/OpenWaterAnalytics/pyswmm).
>
> Not sure about whether USBR or USACE provide access to source code though.

pyHMT2D is something similar to PySWMM. But pyHMT2D itself does not contain the model code. Unfortunately both HEC-RAS and SRH-2D are closed source. So I can't compile them into a library as part of pyHMT2D.
