### 2021-07-28 #3332 @mdbartos
Unfortunately, I do not have access to a physical Windows machine right now, although I am in the process of ordering one through my department. I estimate that it should take a week or three.

If you can provide documentation describing how to install/test pyHMT2D on a cloud instance, that could also work.

If you decide to go with another reviewer, that is also OK.
