### 2021-07-29 #3332 @psu-efd
Again, thanks for your time and effort.

For the use in cloud computing, I am not sure whether you have used AWS, Azure, or Google Cloud. Basically, you need to start a Windows instance (try the free ones) and then remote desktop to the instance. After that, it is the same as your own Windows computer. You can install whatever you like.

From: Matt Bartos ***@***.***>
Sent: Wednesday, July 28, 2021 1:44 AM
Cc: Dr. Xiaofeng Liu, Environmental Fluid Dynamics Group, Penn State University ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: pyHMT2D: Python-based two-dimensional hydraulic modeling tools (#3332)

Unfortunately, I do not have access to a physical Windows machine right now, although I am in the process of ordering one through my department. I estimate that it should take a week or three.

If you can provide documentation describing how to install/test pyHMT2D on a cloud instance, that could also work.

If you decide to go with another reviewer, that is also OK.
