### 2021-08-16 #3332 @kthyng
@dbuscombe-usgs are you interested in reviewing this JOSS software? The review is already underway but we need another reviewer who has access to Windows to complete the process. Would you be able to fill this role? Thanks.
