### 2021-08-16 #3332 @kthyng
@mdbartos You mentioned above that you were able to complete some checkboxes in the review process before needing access to Windows, but none of the boxes are checked above. Were you able to complete some? If so, I would like to preserve that fact by having your review stay in place with the boxes checked that you were able to check along with any notes. If you really weren't able to do much of the review, I can just remove you as a reviewer and remove you from this issue. What do you think?
