### 2021-08-19 #3332 @kthyng
Ok, I think I have everything squared away but please let me know if anything doesn't look correct.

@mdbartos You remain a reviewer, but I will note that not all of the boxes are checked off because you were unable to fully complete the reviewer, not having a Windows machine. Also because you were never able to edit the check boxes, please check the boxes that I checked off to make sure you agree with what is and is not checked.

@dbuscombe-usgs I have added you as a reviewer. I think everything should now be in order for you to check off your own check boxes and pursue the review. Thank you for stepping in to help here!
