### 2021-08-27 e34a017
added function to generate terrain of a composite channel
