### 2021-08-27 7087bd6
added function to generate terrain of a composite channel
