### 2021-09-10 6ab0248
added terrain function for bedform, composite channel; adapted RAS_2D_Data class for HEC-RAS v6, in particular the Manning's n information
