+++ pyHMT2D/Hydraulic_Models_Data/RAS_2D/RAS_2D_Data.py
@@ -608,22 +608,48 @@ class RAS_2D_Data(HydraulicData):
             else:
                 fileBase = str.encode(os.path.dirname(self.hdf_filename) + '/')
 
-            hfManningN = h5py.File(fileBase + self.landcover_layername + b'.hdf', 'r')
 
-            dset = hfManningN['IDs']
+            #The way the land cover and Manning's n information stored in v5 and v6 is different.
+            if self.version == '5.0.7':
+                hfManningN = h5py.File(fileBase + self.landcover_layername + b'.hdf', 'r')
 
-            with dset.astype(np.uint8):
-                IDs = dset[:]
+                dset = hfManningN['IDs']
 
-            ManningN = np.array(hfManningN['ManningsN'])
-            Names = hfManningN['Names']
+                with dset.astype(np.uint8):
+                    IDs = dset[:]
 
-            # print("IDs =", IDs)
-            # print("ManningN =", ManningN)
-            # print("Names =", Names)
+                ManningN = np.array(hfManningN['ManningsN'])
+                Names = hfManningN['Names']
 
-            for i in range(len(IDs)):
-                self.ManningNZones[IDs[i]] = [Names[i], ManningN[i]]
+                # print("IDs =", IDs)
+                # print("ManningN =", ManningN)
+                # print("Names =", Names)
+
+                for i in range(len(IDs)):
+                    self.ManningNZones[IDs[i]] = [Names[i], ManningN[i]]
+
+            elif self.version == '6.0.0':
+                hfManningN = h5py.File(fileBase + self.landcover_filename, 'r')
+
+                dset_raster_map = hfManningN['Raster Map']
+                dset_variables = hfManningN['Variables']
+
+                #with dset_raster_map['ID'].astype(np.uint8):
+                IDs = dset_raster_map['ID']
+
+                ManningN = np.array(dset_variables['ManningsN'])
+                Names = dset_variables['Name']
+
+                # print("IDs =", IDs)
+                # print("ManningN =", ManningN)
+                # print("Names =", Names)
+
+                for i in range(len(IDs)):
+                    self.ManningNZones[IDs[i]] = [Names[i], ManningN[i]]
+
+            else:
+                raise Exception("The version of HEC-RAS that produced this HDF result file is not supported.")
+
 
             if gVerbose: print("self.ManningNZones = ", self.ManningNZones)
 
@@ -631,37 +657,19 @@ class RAS_2D_Data(HydraulicData):
 
         hf.close()
 
-    def modify_ManningsN(self, materialIDs, newManningsNValues, materialNames):
-        """Modify materialID's Manning's n values to new values
-
-        Assume that there is "Land Cover Filename" and "Land Cover Layername" in the result HDF file.
-        If they don't exist (mostly for the case that the whole domain uses the default, constant
-        Manning's n value in ["Geometry"]["2D Flow Areas"][0][1]), just use newManningsNValues[0].
+    def change_ManningsN_v5(self, materialIDs, newManningsNValues, materialNames):
+        """ Change materialID's Mannings n values to new values and save to file (for HEC-RAS v5)
 
         Parameters
         ----------
-        materialIDs : list
-            material ID list
-        newManningsNValues : list
-            new Manning's n values in a list
-        materialNames : list
-            names of the material in a list
-
+        materialIDs
+        newManningsNValues
+        materialNames
 
         Returns
         -------
 
         """
-
-        if gVerbose: print("Modify Manning's n value ...")
-
-        if not isinstance(materialIDs[0], int):
-            print("Material ID has to be an integer. The type of materialID passed in is ", type(materialIDs[0]),
-                  ". Exit.\n")
-
-        if not isinstance(newManningsNValues[0], float):
[…]
