### 2021-11-22 5351ea8
In XMDFC the variables are in the main file but in the XMDF file it is under another variable called Datasets. Therefore in pyHMT2D in SRH_2D_Data.py in the function called readSRHXMDFFile I added one if. also Guid is an extra variable we don't need so I added to to the if.
